examples-from-user-filed-issues.md: clarify multiple corruption handling
Some users may be tempted to only replace one out of several corrupt
objects and then try to see how filter-repo handles it.  The problem is
that having any corrupt object is likely to cause one of fast-export,
filter-repo, or fast-import to fail.  Make it clear that all corrupt
objects need to be replaced before running filter-repo.  (There are some
types of corruption that none of the systems will fail on, and which you
can thus replace one at a time, but we don't want users relying on that
and then coming back to us and asking why it doesn't work.)

Signed-off-by: Elijah Newren <newren@gmail.com>